Finished Archer enemy

+++ Assets/Scripts/Characters/Entity/States/StateDataSO/EntityRangedAttackStateSO.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class EntityRangedAttackStateSO : ScriptableObject
 {
     public GameObject projectile;
-    public float projectileDamage = 10f;
+    public int projectileDamage = 10;
     public float projectileSpeed = 12f;
     public float projectileTravelDistance;
